+++ frontend/src/apis/user/index.ts
@@ -1,8 +1,23 @@
 import https from '@/lib/https';
 
+// 모킹을 사용하지 않으려면 https에 mockApiList에서 /api.auth를 지워주셔야 합니다!
 export const postLogin = async (userId: string) => {
-  const { data } = await https.post(`/api/auth/login`, {
+  const response = await https.post(`/api/auth/login`, {
     userId,
   });
-  return data;
+  const authorization = response.headers.Authorization;
+  if (authorization) {
+    const accessToken = response.headers.authorization.split(' ')[1];
+    localStorage.setItem('access-token', accessToken);
+    localStorage.setItem(
+      'refresh-token',
+      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwidXNlcklkIjoiMDNjNmIwODMtZThkNi00ODhjLWFhODMtMmEwMWIzZjM5ZDAwIiwiaWF0IjoxNTE2MjM5MDIyfQ.iVTdh4kkGh6f6gEZLf9MJPwkjusaXf58z_Tc4ncummw'
+    );
+  }
+  return response;
+};
+
+export const postRefreshToken = async () => {
+  const response = await https.post('/api/auth/refresh');
+  return response;
 };
